requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.3.1=pypi_0
aiosignal=1.4.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asttokens=3.0.0=pypi_0
attrs=25.3.0=pypi_0
ca-certificates=2024.11.26=h06a4308_0
certifi=2025.8.3=pypi_0
charset-normalizer=3.4.3=pypi_0
click=8.1.8=pypi_0
cloudpickle=3.1.1=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
dm-control=1.0.32=pypi_0
dm-env=1.6=pypi_0
dm-tree=0.1.8=pypi_0
et-xmlfile=2.0.0=pypi_0
etils=1.5.2=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.1.0=pypi_0
farama-notifications=0.0.4=pypi_0
filelock=3.18.0=pypi_0
fonttools=4.55.3=pypi_0
frozenlist=1.7.0=pypi_0
fsspec=2025.5.1=pypi_0
glfw=2.9.0=pypi_0
gymnasium=1.1.1=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.7.0=pypi_0
importlib-resources=6.4.5=pypi_0
ipython=8.18.1=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.5.0=pypi_0
jsonschema=4.25.1=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
kiwisolver=1.4.7=pypi_0
labmaze=1.0.6=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lxml=6.0.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.1=pypi_0
mujoco=3.3.5=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
numpy=1.26.1=pypi_0
nvidia-cublas-cu12=12.6.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.6.80=pypi_0
nvidia-cuda-nvrtc-cu12=12.6.77=pypi_0
nvidia-cuda-runtime-cu12=12.6.77=pypi_0
nvidia-cudnn-cu12=9.5.1.17=pypi_0
nvidia-cufft-cu12=11.3.0.4=pypi_0
nvidia-cufile-cu12=1.11.1.6=pypi_0
nvidia-curand-cu12=10.3.7.77=pypi_0
nvidia-cusolver-cu12=11.7.1.2=pypi_0
nvidia-cusparse-cu12=12.5.4.2=pypi_0
nvidia-cusparselt-cu12=0.6.3=pypi_0
nvidia-nccl-cu12=2.26.2=pypi_0
nvidia-nvjitlink-cu12=12.6.85=pypi_0
nvidia-nvtx-cu12=12.6.77=pypi_0
omegaconf=2.3.0=pypi_0
opencv-contrib-python=4.11.0.86=pypi_0
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py39h06a4308_0
prompt-toolkit=3.0.48=pypi_0
protobuf=6.32.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=18.1.0=pypi_0
pygments=2.18.0=pypi_0
pymdptoolbox=4.0b3=pypi_0
pyopengl=3.1.10=pypi_0
pyparsing=3.2.0=pypi_0
python=3.9.21=he870216_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
ray=2.10.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
requests=2.32.5=pypi_0
rich=14.1.0=pypi_0
rpds-py=0.27.0=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=pypi_0
setuptools=75.1.0=py39h06a4308_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.13.3=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h39e8969_0
torch=2.7.0=pypi_0
torchaudio=2.7.0=pypi_0
torchvision=0.22.0=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
triton=3.3.0=pypi_0
typer=0.16.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.5.0=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.44.0=py39h06a4308_0
xz=5.4.6=h5eee18b_1
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
